refactor: Command와 Query 역할 분리를 위한 Facade 리팩토링 (#94)
