+++ src/main/kotlin/kr/wooco/woocobe/place/ui/web/facade/PlaceReviewQueryFacade.kt
@@ -1,7 +1,6 @@
 package kr.wooco.woocobe.place.ui.web.facade
 
 import kr.wooco.woocobe.place.domain.usecase.AddPlaceReviewUseCase
-import kr.wooco.woocobe.place.domain.usecase.DeletePlaceReviewInput
 import kr.wooco.woocobe.place.domain.usecase.DeletePlaceReviewUseCase
 import kr.wooco.woocobe.place.domain.usecase.GetAllMyPlaceReviewUseCase
 import kr.wooco.woocobe.place.domain.usecase.GetAllPlaceReviewInput
@@ -10,9 +9,6 @@ import kr.wooco.woocobe.place.domain.usecase.GetPlaceReviewInput
 import kr.wooco.woocobe.place.domain.usecase.GetPlaceReviewUseCase
 import kr.wooco.woocobe.place.domain.usecase.GetUserAllPlaceReviewInput
 import kr.wooco.woocobe.place.domain.usecase.UpdatePlaceReviewUseCase
-import kr.wooco.woocobe.place.ui.web.controller.request.CreatePlaceReviewRequest
-import kr.wooco.woocobe.place.ui.web.controller.request.UpdatePlaceReviewRequest
-import kr.wooco.woocobe.place.ui.web.controller.response.CreatePlaceReviewResponse
 import kr.wooco.woocobe.place.ui.web.controller.response.PlaceReviewDetailsResponse
 import kr.wooco.woocobe.user.domain.usecase.GetAllUserInput
 import kr.wooco.woocobe.user.domain.usecase.GetAllUserUseCase
@@ -31,27 +27,6 @@ class PlaceReviewQueryFacade(
     private val getUserUseCase: GetUserUseCase,
     private val getAllUserUseCase: GetAllUserUseCase,
 ) {
-    fun createPlaceReview(
-        userId: Long,
-        placeId: Long,
-        request: CreatePlaceReviewRequest,
-    ): CreatePlaceReviewResponse {
-        val addPlaceReviewResult = addPlaceReviewUseCase.execute(request.toCommand(userId = userId, placeId = placeId))
-
-        return CreatePlaceReviewResponse(addPlaceReviewResult.placeReviewId)
-    }
-
-    fun updatePlaceReview(
-        userId: Long,
-        placeReviewId: Long,
-        request: UpdatePlaceReviewRequest,
-    ) = updatePlaceReviewUseCase.execute(request.toCommand(userId = userId, placeReviewId = placeReviewId))
-
-    fun deletePlaceReview(
-        userId: Long,
-        placeReviewId: Long,
-    ) = deletePlaceReviewUseCase.execute(DeletePlaceReviewInput(userId = userId, placeReviewId = placeReviewId))
-
     fun getPlaceReviewDetail(placeReviewId: Long): PlaceReviewDetailsResponse {
         val getPlaceReviewResult = getPlaceReviewUseCase.execute(GetPlaceReviewInput(placeReviewId = placeReviewId))
